We now support Python3, but there are some issues (see description)
The 2to3 build works smoothly, the only problem is, if we run
``python setup.py test``, the "stock" smartrpyc will still be in PYTHONPATH
and get loaded instead of the one in site-packages (or build/lib).

A solution to run the tests is to::

  cd build/lib
  python -m unittest discover -v smartrpyc.tests

(Anyways, the installed version seems to be working just fine).